Update launcher to support up-coming new texture_replacement format (#494)
Didn't bother trying to make this backwards compatible, texture packs
will need to update, and users will need to update the launcher and
tooling to the latest version. It would be too annoying to make this
fully non-breaking (ie. people update their tooling but not the launcher
or vise-versa). Not worth the effort.

Release once 0.2.13 is out, this has been tested end-to-end:
![Screenshot 2024-05-30
212348](https://github.com/open-goal/launcher/assets/13153231/2765059a-c414-4f1b-bbfb-0a8470907bac)
![Screenshot 2024-05-30
212539](https://github.com/open-goal/launcher/assets/13153231/62346158-f498-4793-8cb8-c2529722e495)
![Screenshot 2024-05-30
212402](https://github.com/open-goal/launcher/assets/13153231/f6df5a8f-d807-4960-b10a-362837fed514)