frontend: simplify the version pagination regex (#593)
Fixes #586 

Depending on your javascript engine, you may or may not have support for
lookaheads/lookbehinds. I took this regex from github's examples without
thinking -- probably fine in all versions of `node` but of course some
platforms it's not going to work.

Simplify it to be a more typical regex.