fixes for tauri v2 (#666)
misc changes i find while testing.

tauri changed the platform() function in v2, it's no longer async so we
dont need to await. this lets us remove it from the onMount function and
use it in the top level script. the platform() function also returns
"windows", "macos", or "linux" now in lieu of "win32" and "darwin". this
allowed me to simplify the logic in mods.ts.

added /gen folder to the gitignore

did a full pass through with cargo clippy to adhere to rust best practices