Refactor vcc check (#710)
These changes are directed towards reducing the number of times we check
the required VCC version by shifting to only run the check to startup
(on splash). This change uses a svelte writable store to contain the
result of the VCC check and that store is accessed by all relevant
components.

The motivation for this change is to prevent excessive logs of the VCC
version. This change is part of a broader series of refactors I'm
working on which will hopefully result in cleaner logs.

Additionally, there are a few misc. changes I've made along the way. In
Tauri v2 `type()` and `arch()` are sync which means we can clean up the
code a bit.