ci: Ensure download links in updater metadata file use version-specific links (#719)
This adjusts the automatically spit out tauri metadata to always be
version specific downloads. Most platforms dont have this problem
because the artifact itself has the version number, the macOS artifacts
though do not and only the arm one elects to use the version link?
Weird, likely a tauri-action issue.

Either way this is the better way to do it, download links point to a
more static target instead of a moving one