Allow the main window to be resized and allow linux users to change their installation dir again (#856)
The CSS on the app is properly responsive, the window no longer has to
be a fixed size. This should allow it to nicely expand to screens like
the steamdeck where the application takes up the entire screen (in game
mode) instead of having scaling compromises.

Also fixes the CSS around the last progress bar/icon where there would
be no gap between the bar and the icon.